+++ hubdle-settings-gradle-plugin/main/kotlin/com/javiersc/hubdle/settings/HubdleSettingsPlugin.kt
@@ -49,6 +49,7 @@ private fun Settings.configureRepositories() {
         repositories {
             mavenCentral()
             google()
+            gradlePluginPortal()
             maven(url = "https://maven.pkg.jetbrains.space/public/p/compose/dev")
         }
     }