+++ CHANGELOG.md
@@ -21,6 +21,11 @@
 
 - `org.jetbrains.kotlin:kotlin-serialization -> 1.7.20`
 - `org.jetbrains.kotlin:kotlin-gradle-plugin -> 1.7.20`
+- `org.jetbrains.kotlin:kotlin-test-testng -> 1.7.20`
+- `org.jetbrains.kotlin:kotlin-test-junit5 -> 1.7.20`
+- `org.jetbrains.kotlin:kotlin-test-junit -> 1.7.20`
+- `org.jetbrains.kotlin:kotlin-test-annotations-common -> 1.7.20`
+- `org.jetbrains.kotlin:kotlin-test -> 1.7.20`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.38`
 
 ## [0.2.0-alpha.38] - 2022-09-29