+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Updated
 
+- `gradle/gradle-build-action -> v2.3.2`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.2.0-alpha.39`
 
 ## [0.2.0-alpha.39] - 2022-09-29